requirements.txt
gunicorn
django-heroku
django-cors-headers
pandas
boto3
numpy
